requirements.txt
Django==6.0.4
daphne==4.2.1
channels==4.3.2
channels-redis==4.2.1
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.8
graphql-relay==3.2.0
django-graphql-jwt==0.4.0
django-cors-headers==4.9.0
PyJWT==2.12.1
kafka-python==2.3.1
asgiref==3.11.1
Twisted==25.5.0
autobahn==25.12.2
txaio==25.12.2
pyOpenSSL==26.1.0
attrs==26.1.0
service-identity==24.2.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
whitenoise==6.9.0
dj-database-url==2.3.0
